Course Proposal/requirements.txt
python-dotenv==1.0.1
google-generativeai==0.8.3
Flask==3.1.0

